Update 2025-09-23-staton.md
I added a quote in relation to the topic to back up the point in a stronger way. I think a quote can give this discussion more depth. A few more could benefit the post as well, but textual evidence is stronger than just saying it happened or was said.